requirements.txt
streamlit==1.46.1
requests==2.32.4
python-dotenv==1.1.0
markdown==3.8.2
xhtml2pdf==0.2.17